+++ www/go/core/views/extjs3/form/Chips.js
@@ -190,6 +190,10 @@ go.form.Chips = Ext.extend(Ext.Container, {
 		if(!values) {
 			values = this.map ? {} : [];
 		}
+
+		if(this.comboBox) {
+			this.comboBox.clearInvalid();
+		}
 		
 		if(this.entityStore) {	
 			var ids = this.map ? Object.keys(values) : values;
@@ -232,15 +236,13 @@ go.form.Chips = Ext.extend(Ext.Container, {
 	},
 	markInvalid: function (msg) {		
 		if(this.comboBox) {
-			this.comboBox.getEl().addClass('x-form-invalid');
+			this.comboBox.markInvalid(msg);
 		}
-		Ext.form.MessageTargets.qtip.mark(	this.comboBox, msg);
 	},
 	clearInvalid: function () {
 		if(this.comboBox) {
-			this.comboBox.getEl().removeClass('x-form-invalid');
+			this.comboBox.clearInvalid();
 		}
-		Ext.form.MessageTargets.qtip.clear(this.comboBox);
 	},
 	createComboBox: function () {
 		if(this.store) {